Innfører malbiblioteks-klasse, for å få isolert dette ansvaret frå alt anna TemplateResource har
